python/requirements.txt
Flask
Flask-CORS
requests
pandas
numpy
scikit-learn
joblib
kubernetes
openai
langchain
langchain-community
langchain-experimental
langchain-openai
langgraph
pydantic
statsmodels
scipy
PyYAML
tavily-python
python-dotenv
pytest
pytest-asyncio
psutil